# Enable grpc encoder with pooling