<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

Benchmark distributed read: (The test is added in PR https://github.com/buildbuddy-io/buildbuddy/pull/5487)
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                    │ /home/lulu/benchmark-distributed-read-bs.log │ /home/lulu/benchmark-distributed-read-marshal-pool.log │
                    │                    sec/op                    │             sec/op               vs base               │
Read/digest128B-24                                     171.9µ ± 3%                       173.0µ ± 4%       ~ (p=0.796 n=10)
Read/digest16KiB-24                                    237.6µ ± 5%                       236.3µ ± 5%       ~ (p=0.165 n=10)
Read/digest16MiB-24                                    31.01m ± 5%                       29.35m ± 4%  -5.35% (p=0.005 n=10)
geomean                                                1.082m                            1.063m       -1.79%

                    │ /home/lulu/benchmark-distributed-read-bs.log │ /home/lulu/benchmark-distributed-read-marshal-pool.log │
                    │                     B/op                     │              B/op               vs base                │
Read/digest128B-24                                    52.35Ki ± 2%                    52.43Ki ±  2%        ~ (p=0.631 n=10)
Read/digest16KiB-24                                   86.58Ki ± 2%                    69.04Ki ±  2%  -20.26% (p=0.000 n=10)
Read/digest16MiB-24                                   38.93Mi ± 4%                    32.62Mi ± 12%  -16.21% (p=0.000 n=10)
geomean                                               565.3Ki                         494.5Ki        -12.53%

                    │ /home/lulu/benchmark-distributed-read-bs.log │ /home/lulu/benchmark-distributed-read-marshal-pool.log │
                    │                  allocs/op                   │            allocs/op             vs base               │
Read/digest128B-24                                      354.0 ± 0%                        356.0 ± 0%  +0.56% (p=0.000 n=10)
Read/digest16KiB-24                                     361.5 ± 0%                        364.0 ± 1%  +0.69% (p=0.000 n=10)
Read/digest16MiB-24                                    2.002k ± 5%                       2.018k ± 7%       ~ (p=0.481 n=10)
geomean                                                 635.1                             639.5       +0.69%
```